Miscellaneous cleanup and individual file loading and saving
