+++ Shared/Extensions/CGSizeExtensions.swift
@@ -31,22 +31,5 @@ extension CGSize {
         let boardRatio = boardDimensions.height / boardDimensions.width
         
         return (abs(verticalRatio-boardRatio) < abs(horizontalRatio-boardRatio)) ? .vertical : .horizontal
-        }
-        
-//        (boardDimensions.height / boardDimensions.width < self.height / (self.width-withHorizontalPlayerinfoWidth) ? .vertical : .horizontal)
-    
-//    func fitBox(boardDimensions: CGSize, withHorizontalOffsets hOffsets: CGSize, withVerticalOffsets vOffsets: CGSize) -> CGSize {
-//        if (reccomendedLayoutOrinetationForShape(boardDimensions: boardDimensions, withHorizontalOffsets: hOffsets, withVerticalOffsets: vOffsets) == .horizontal) {
-//            let horizontalRatio = (self.height - hOffsets.height) / (self.width - hOffsets.width)
-//            return CGSize(width: <#T##CGFloat#>, height: <#T##CGFloat#>)
-//        } else {
-//            let verticalRatio = (self.height - vOffsets.height) / (self.width - vOffsets.width)
-//
-//        }
-//    }
-    
-//    func reccomendedScrolling(boardDimensions: CGSize, withHorizontalOffsets hOffsets: CGSize, withVerticalOffsets vOffsets: CGSize, threshold: CGFloat) -> Bool {
-//        let fit = fitBox(size: boardDimensions)
-//        return fit.width <= threshold || fit.height <= threshold
-//    }
+    }
 }
