+++ Shared/Models/Chess-Board/ChessBoard1v1v1v1.swift
@@ -196,9 +196,7 @@ struct ChessBoard1v1v1v1: TraditionalRulesChessBoard, Codable {
         enPassentPositions = [PlayerID : Position]()
     }
     
-    
-    
-    //MARK: TODO!!!
+    //MARK: TODO get position name
     func getPositionName(_ position: Position) -> String {
         "TODO!"
     }
