+++ Shared/Models/Players/ComputerPlayer.swift
@@ -25,7 +25,6 @@ struct ComputerPlayer: Player {
         DispatchQueue.main.asyncAfter(deadline: .now() + 0.2) {
             makeArbitraryMove(withBoard: chessBoard)
         }
-        
     }
     
     private func makeArbitraryMove(withBoard chessBoard: ChessBoard) {
