+++ Shared/View-Model/ChessGame.swift
@@ -7,30 +7,31 @@
 
 import SwiftUI
 
-struct RenderablePiece: Identifiable {
-    let piece: Piece
-    var id: Int { piece.id }
-    var position: Position
-}
-
 class ChessGame: ObservableObject, CustomStringConvertible, Identifiable, Codable {
+    @EnvironmentObject var settings: AppSettings
+    let ANIMATION_DURATION = 0.1
+
     // MARK: - Board Identity
     let gameType: ChessGameType
+    @Published private(set) var chessBoard: ChessBoard
+    @Published  var gameState: GameState
+    
     private(set) var id: UUID
     private(set) var players: [Player]
     private(set) var history: [ChessBoard]
     private(set) var playersInCheck: Set<PlayerID>
-    
-    @Published private(set) var chessBoard: ChessBoard
-    @Published  var gameState: GameState
-    @EnvironmentObject var settings: AppSettings
-    let animationDuration = 0.1
-    
+    var hasDisplayedGameOver = false
+
+    // MARK: Temporary State
     @Published var pieceShowingPromotionView: Int? = nil
     @Published var promotionPieceTypes: [PieceType] = []
+    @Published private(set) var potentialMoveDestinations: Set<Position>
+    @Published private(set) var warningPositions: Set<Position>
+    @Published private(set) var lastMoves: Set<Position>
     var potentialPromotionFromPosition: Position? = nil
     var potentialPromotionDestination: Position? = nil
-
+     
+    // MARK: Computed Properties
     var currentPlayer: Player? {
         if let id = gameState.getCurrentPlayer() {
             return playerWithID(id: id)
@@ -38,24 +39,28 @@ class ChessGame: ObservableObject, CustomStringConvertible, Identifiable, Codabl
             return nil
         }
     }
+    
     var nextPlayer: Player? {
         if let nextPlayerIndex = currentPlayer?.nextPlayer.index {
             return players[nextPlayerIndex]
         } else { return nil }
     }
+    
     var playersStillInGame: [PlayerID] {
         var identities: [PlayerID] = []
         players.filter { !$0.hasBeenEliminated }
-            .forEach { player in identities.append(player.identity)}
+        .forEach { player in identities.append(player.identity)}
         return identities
     }
+    
     func playersInTeam(_ team: TeamID) -> [Player] {
         players.filter({player in player.team == team})
     }
+    
     func playerWithID(id: PlayerID) -> Player? {
         return players.filter(){player in player.identity == id}.first
     }
-
+    
     var renderedChessPieces: [RenderablePiece] {
         var pieces = [RenderablePiece]()
         chessBoard.board.keys.forEach { position in
@@ -64,6 +69,26 @@ class ChessGame: ObservableObject, CustomStringConvertible, Identifiable, Codabl
         return pieces
     }
     
+    var description: String {
+        var description = ""
+        var teams: [String] = []
+        var teamIDs: [TeamID] = []
+        for player in players {
+            if let index = teamIDs.firstIndex(where: { teamID in return teamID == player.team }) {
+                teams[index] += ", \(player.name)"
+            } else {
+                teamIDs.append(player.team)
+                teams.append(player.name)
+            }
+        }
+        
+        for teamIndex in 0..<teams.count {
+            description += teamIndex == 0 ? teams[teamIndex] : " vs \(teams[teamIndex])"
+        }
+        return description
+    }
+    
+    // MARK: - Initializer
     init(type: ChessGameType, playerBuilders: [PlayerBuilder], id: UUID = UUID()) {
         self.id = id
         history = []
@@ -87,11 +112,10 @@ class ChessGame: ObservableObject, CustomStringConvertible, Identifiable, Codabl
         players = []
         let playerCount = playerBuilders.count
         var i = 0
-
+        
         for builder in playerBuilders {
             assert(builder.team == gameType.expectedTeamIDs[i], "Incorrect team ID's for game type \(gameType) for game: \(self)")
             i+=1
-
             let nextPlayer = (players.count == playerCount-1) ? 0 : players.count+1
             let previousPlayer = players.isEmpty ? playerCount-1 : players.count-1
             let responseHandler: PlayerResponseHandler = {move in self.handleMove(move)}
@@ -102,7 +126,8 @@ class ChessGame: ObservableObject, CustomStringConvertible, Identifiable, Codabl
         self.players.first!.startMove(withBoard: chessBoard)
         gameState = .waitingOnPlayer(player: self.players.first!.identity)
     }
-    
+ 
+    // MARK: - Codablity
     enum ChessGameCodingKeys: String, CodingKey {
         case gameType
         case id
@@ -111,77 +136,48 @@ class ChessGame: ObservableObject, CustomStringConvertible, Identifiable, Codabl
         case playersInCheck
         case chessBoard
         case gameState
-        case settings
         case lastMoves
         case warningPositions
+        case hasDisplayedGameOver
     }
     
     required init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: ChessGameCodingKeys.self)
-        
         gameType = try container.decode(ChessGameType.self, forKey: .gameType)
         id = try container.decode(UUID.self, forKey: .id)
-        
         var historyContainer = try container.nestedUnkeyedContainer(forKey: .history)
         history = try ChessBoardCoder.decodeChessBoardList(container: &historyContainer, gameType: gameType)
-        
         players = []
         gameState = .paused
-       
         playersInCheck = try container.decode(Set<PlayerID>.self, forKey: .playersInCheck)
         chessBoard = try ChessBoardCoder.decodeKeyed(container: container, key: .chessBoard, gameType: gameType)
-//        settings
         lastMoves = try container.decode(Set<Position>.self, forKey: .lastMoves)
         warningPositions = try container.decode(Set<Position>.self, forKey: .warningPositions)
         potentialMoveDestinations = []
-        
         var playersContainer = try container.nestedUnkeyedContainer(forKey: .players)
         let responseHandler: PlayerResponseHandler = {move in self.handleMove(move)}
         players = try PlayerCoder.decodePlayerList(fromContainer: &playersContainer, handler: responseHandler)
         gameState = try container.decode(GameState.self, forKey: .gameState)
+        hasDisplayedGameOver = try container.decode(Bool.self, forKey: .hasDisplayedGameOver)
     }
-
+    
     func encode(to encoder: Encoder) throws {
         var container = encoder.container(keyedBy: ChessGameCodingKeys.self)
-        
         try container.encode(gameType, forKey: .gameType)
         try container.encode(id, forKey: .id)
-        
         var playerContainer = container.nestedUnkeyedContainer(forKey: .players)
         try PlayerCoder.encodePlayerList(container: &playerContainer, players: players)
-        
         var historyContainer = container.nestedUnkeyedContainer(forKey: .history)
         try ChessBoardCoder.encodeChessBoardList(container: &historyContainer, boards: history)
-        
         try container.encode(playersInCheck, forKey: .playersInCheck)
         try ChessBoardCoder.encodeKeyed(container: &container, key: .chessBoard, board: chessBoard)
         try container.encode(gameState, forKey: .gameState)
-        //            try container.encode(settings, forKey: .settings)
         try container.encode(lastMoves, forKey: .lastMoves)
         try container.encode(warningPositions, forKey: .warningPositions)
-    }
-    
-    var description: String {
-        var description = ""
-        var teams: [String] = []
-        var teamIDs: [TeamID] = []
-        for player in players {
-            if let index = teamIDs.firstIndex(where: { teamID in return teamID == player.team }) {
-                teams[index] += ", \(player.name)"
-            } else {
-                teamIDs.append(player.team)
-                teams.append(player.name)
-            }
-        }
-        
-        for teamIndex in 0..<teams.count {
-            description += teamIndex == 0 ? teams[teamIndex] : " vs \(teams[teamIndex])"
-        }
-        return description
+        try container.encode(hasDisplayedGameOver, forKey: .hasDisplayedGameOver)
     }
     
     // MARK: - View Events
-    
     private var potentialMovesForCurrentPiece: [Move] = []
     @Published private(set) var userFocusedPosition: Position? {
         didSet {
@@ -194,17 +190,7 @@ class ChessGame: ObservableObject, CustomStringConvertible, Identifiable, Codabl
             }
         }
     }
-    @Published private(set) var potentialMoveDestinations: Set<Position>
-    @Published private(set) var warningPositions: Set<Position>
-    @Published private(set) var lastMoves: Set<Position>
-    
-    enum SelectionType {
-        case userFocus
-        case potentialMove
-        case warning
-        case lastMove
-    }
-    
+   
     func userTappedPosition(_ position: Position) {
         let animation = Animation.linear(duration: 0.1)
         if gameState.isWaitingOnUserToMakeMove(game: self) {
@@ -240,7 +226,6 @@ class ChessGame: ObservableObject, CustomStringConvertible, Identifiable, Codabl
                 }
             }
         }
-            
         clearHighlights()
     }
     
@@ -248,18 +233,16 @@ class ChessGame: ObservableObject, CustomStringConvertible, Identifiable, Codabl
         pieceShowingPromotionView = nil
         promotionPieceTypes.removeAll()
         objectWillChange.send()
-        
         if let type = pieceType {
             let promotionMoves = chessBoard.getMoves(from: potentialPromotionFromPosition!).filter({ $0.promotionType == type && $0.primaryDestination == potentialPromotionDestination!})
             let onDevicePlayer = currentPlayer as! OnDevicePlayer
             let move = promotionMoves.first!
             onDevicePlayer.handleOnDeviceMove(move)
-
         }
     }
     
     func clearHighlights() {
-        withAnimation(.linear(duration: animationDuration)) {
+        withAnimation(.linear(duration: ANIMATION_DURATION)) {
             potentialMoveDestinations.removeAll()
             potentialMovesForCurrentPiece.removeAll()
             userFocusedPosition = nil
@@ -270,7 +253,7 @@ class ChessGame: ObservableObject, CustomStringConvertible, Identifiable, Codabl
         if case .waitingOnPlayer(let currentPlayer) = gameState,
            let piece = chessBoard.board[position],
            piece.player == currentPlayer {
-            withAnimation(.linear(duration: animationDuration)) {
+            withAnimation(.linear(duration: ANIMATION_DURATION)) {
                 userFocusedPosition = position
                 potentialMovesForCurrentPiece = chessBoard.getMoves(from: position)
                 potentialMoveDestinations.removeAll()
@@ -286,112 +269,27 @@ class ChessGame: ObservableObject, CustomStringConvertible, Identifiable, Codabl
                 chessBoard = chessBoard.doMove(move)
             }
             
-            withAnimation(.linear(duration: animationDuration)) {
+            withAnimation(.linear(duration: ANIMATION_DURATION)) {
                 if let oldLastMove = firstPlayer.lastMove {
                     lastMoves.remove(oldLastMove.primaryDestination)
                 }
                 lastMoves.insert(move.primaryDestination)
                 players[firstPlayer.index].lastMove = move
-                
                 updateGameState()
             }
-
+            
             if let nextPlayer = currentPlayer {
                 nextPlayer.startMove(withBoard: chessBoard)
             }
             
             history.append(chessBoard)
             objectWillChange.send()
-            ChessGameStore.saveGames()
+            ChessGameStore.instance.saveGame(game: self)
         }
-
     }
     
     // MARK: - States
-    
-    enum GameState: Codable {
-        case paused
-        case waitingOnPlayer(player: PlayerID)
-        case endedVictory(forTeam: TeamID)
-        case endedStalemate
         
-        
-        func isWaitingOnUserToMakeMove(game: ChessGame) -> Bool {
-            if let currentPlayer = getCurrentPlayer() {
-                return game.playerWithID(id: currentPlayer)?.type == .onDevice
-            } else {
-                return false
-            }
-        }
-        
-        func isWaitingOnComputer(game: ChessGame) -> Bool {
-            if let currentPlayer = getCurrentPlayer() {
-                return game.playerWithID(id: currentPlayer)?.type == .computer
-            } else {
-                return false
-            }
-        }
-        
-        func getCurrentPlayer() -> PlayerID? {
-            if case .waitingOnPlayer(let currentPlayer) = self {
-                return currentPlayer
-                
-            } else {
-                return nil
-            }
-        }
-        
-        var gameOver: Bool {
-            switch self {
-            case .endedStalemate:
-                return true
-            case .endedVictory(_):
-                return true
-            default:
-                return false
-            }
-        }
-        
-        // Codability
-        enum GameStateCodingKeys: CodingKey {
-            case state, waitingPlayer, winningTeam
-        }
-        
-        init(from decoder: Decoder) throws {
-            let container = try decoder.container(keyedBy: GameStateCodingKeys.self)
-            
-            let stateIndex = try container.decode(Int.self, forKey: .state)
-            switch stateIndex {
-            case 0:
-                self = .paused
-            case 1:
-                self = .waitingOnPlayer(player: try container.decode(PlayerID.self, forKey: .waitingPlayer))
-            case 2:
-                self = .endedVictory(forTeam: try container.decode(TeamID.self, forKey: .winningTeam))
-            case 3:
-                self = .endedStalemate
-            default:
-                throw DecodingError.dataCorruptedError(forKey: .state, in: container, debugDescription: "Unrecognized game state")
-            }
-        }
-        
-        func encode(to encoder: Encoder) throws {
-            var container = encoder.container(keyedBy: GameStateCodingKeys.self)
-            switch self {
-            case .paused:
-                try container.encode(0, forKey: .state)
-            case .waitingOnPlayer(let player):
-                try container.encode(1, forKey: .state)
-                try container.encode(player, forKey: .waitingPlayer)
-            case .endedVictory(let forTeam):
-                try container.encode(2, forKey: .state)
-                try container.encode(forTeam, forKey: .winningTeam)
-            case .endedStalemate:
-                try container.encode(3, forKey: .state)
-            }
-        }
-    }
-    
     func updateGameState() {
         if let next = nextPlayer {
             ChessGameStore.instance.gameWillChange()
@@ -408,13 +306,13 @@ class ChessGame: ObservableObject, CustomStringConvertible, Identifiable, Codabl
                 gameState = .waitingOnPlayer(player: next.identity)
             } else {
                 if playersInCheck.contains(next.identity) {
-//                    eliminatePlayer(playerID: next.identity)
+                    //                    eliminatePlayer(playerID: next.identity)
                     eliminateTeam(teamID: next.team) // Player eliminated when can't make a move and in check
                 } else {
                     if players.count == 2 {
                         gameState = .endedStalemate // 1v1 Games stalemated when player can't make move and isn't in check
                     } else {
-//                        eliminatePlayer(playerID: next.identity)
+                        //                        eliminatePlayer(playerID: next.identity)
                         eliminateTeam(teamID: next.team) // Non 1v1 games eliminate player when they can't make a move
                     }
                 }
@@ -447,9 +345,5 @@ class ChessGame: ObservableObject, CustomStringConvertible, Identifiable, Codabl
             }
         }
     }
-    
-    var hasDisplayedGameOver = false
-    
-   
 }
