+++ Shared/Views/GameViews/DuelGameView.swift
@@ -18,13 +18,6 @@ struct DuelGameView: View {
     
     var horizontalOffsets: CGSize {
         return CGSize(width: 150, height: 0)
-//        var maxLength = 5
-//        for player in boardView.chessGame.players {
-//            maxLength = max(maxLength, player.name.count)
-//        }
-//        maxLength = min(maxLength*8+20+20, 150)
-//        print(maxLength)
-//        return CGSize(width: maxLength, height: 0)
     }
     var verticalOffsets: CGSize {
         CGSize(width: 0, height: 25+10)
@@ -72,8 +65,5 @@ struct DuelGameView: View {
             boardView.layoutPriority(1)
             Spacer()
         }
-        
     }
 }
-
-
